backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
python-dotenv
openai
pandas
openpyxl
requests
beautifulsoup4
selenium
sqlite-vec
hdbscan
numpy
scikit-learn